+++ pair_homekit.c
@@ -739,7 +739,7 @@ tlv_debug(const tlv_values_t *values)
 #endif
 
 static tlv_values_t *
-response_process(const uint8_t *data, uint32_t data_len, const char **errmsg)
+response_process(const uint8_t *data, size_t data_len, const char **errmsg)
 {
   tlv_values_t *response;
   tlv_t *error;
@@ -1081,7 +1081,7 @@ pair_setup_free(struct pair_setup_context *sctx)
 }
 
 static uint8_t *
-pair_setup_request1(uint32_t *len, struct pair_setup_context *sctx)
+pair_setup_request1(size_t *len, struct pair_setup_context *sctx)
 {
   tlv_values_t *request;
   uint8_t *data;
@@ -1138,7 +1138,7 @@ pair_setup_request1(uint32_t *len, struct pair_setup_context *sctx)
 }
 
 static uint8_t *
-pair_setup_request2(uint32_t *len, struct pair_setup_context *sctx)
+pair_setup_request2(size_t *len, struct pair_setup_context *sctx)
 {
   tlv_values_t *request;
   uint8_t *data;
@@ -1179,7 +1179,7 @@ pair_setup_request2(uint32_t *len, struct pair_setup_context *sctx)
 }
 
 static uint8_t *
-pair_setup_request3(uint32_t *len, struct pair_setup_context *sctx)
+pair_setup_request3(size_t *len, struct pair_setup_context *sctx)
 {
   tlv_values_t *request;
   uint8_t *data;
@@ -1282,7 +1282,7 @@ pair_setup_request3(uint32_t *len, struct pair_setup_context *sctx)
 }
 
 static int
-pair_setup_response1(struct pair_setup_context *sctx, const uint8_t *data, uint32_t data_len)
+pair_setup_response1(struct pair_setup_context *sctx, const uint8_t *data, size_t data_len)
 {
   tlv_values_t *response;
   tlv_t *pk;
@@ -1319,7 +1319,7 @@ pair_setup_response1(struct pair_setup_context *sctx, const uint8_t *data, uint3
 }
 
 static int
-pair_setup_response2(struct pair_setup_context *sctx, const uint8_t *data, uint32_t data_len)
+pair_setup_response2(struct pair_setup_context *sctx, const uint8_t *data, size_t data_len)
 {
   tlv_values_t *response;
   tlv_t *proof;
@@ -1362,7 +1362,7 @@ pair_setup_response2(struct pair_setup_context *sctx, const uint8_t *data, uint3
 }
 
 static int
-pair_setup_response3(struct pair_setup_context *sctx, const uint8_t *data, uint32_t data_len)
+pair_setup_response3(struct pair_setup_context *sctx, const uint8_t *data, size_t data_len)
 {
   tlv_values_t *response;
   tlv_t *encrypted_data;
@@ -1473,7 +1473,7 @@ pair_setup_result(const uint8_t **key, size_t *key_len, struct pair_setup_contex
 }
 
 static uint8_t *
-pair_verify_request1(uint32_t *len, struct pair_verify_context *vctx)
+pair_verify_request1(size_t *len, struct pair_verify_context *vctx)
 {
   const uint8_t basepoint[32] = {9};
   tlv_values_t *request;
@@ -1514,7 +1514,7 @@ pair_verify_request1(uint32_t *len, struct pair_verify_context *vctx)
 }
 
 static uint8_t *
-pair_verify_request2(uint32_t *len, struct pair_verify_context *vctx)
+pair_verify_request2(size_t *len, struct pair_verify_context *vctx)
 {
   tlv_values_t *request;
   uint8_t *data;
@@ -1584,7 +1584,7 @@ pair_verify_request2(uint32_t *len, struct pair_verify_context *vctx)
 }
 
 static int
-pair_verify_response1(struct pair_verify_context *vctx, const uint8_t *data, uint32_t data_len)
+pair_verify_response1(struct pair_verify_context *vctx, const uint8_t *data, size_t data_len)
 {
   tlv_values_t *response;
   tlv_t *encrypted_data;
@@ -1671,7 +1671,7 @@ pair_verify_response1(struct pair_verify_context *vctx, const uint8_t *data, uin
 }
 
 static int
-pair_verify_response2(struct pair_verify_context *vctx, const uint8_t *data, uint32_t data_len)
+pair_verify_response2(struct pair_verify_context *vctx, const uint8_t *data, size_t data_len)
 {
   // TODO actually check response
   return 0;
